Rename no-border hoc into index.js
