+++ src/components/20-molecules/cookie-disclaimer/react/story.jsx
@@ -4,7 +4,7 @@ import { storiesOf } from '@storybook/html';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import AXACookieDisclaimerReact from './AXACookieDisclaimerReact';
-import withNoBorder from '../../../../../.storybook/addons/no-border/no-border';
+import withNoBorder from '../../../../../.storybook/addons/no-border';
 import AXALinkReact from './AXALinkReact';
 import Readme from '../README.md';
 
@@ -38,5 +38,14 @@ storiesOf('Molecules/Cookie disclaimer/React', module)
       </AXACookieDisclaimerReact>,
       div
     );
+    const warning = document.createElement('div');
+    warning.innerHTML = `
+      <!-- This is code only for the demo -->
+      <br/>
+      <div style="border: 1px solid red; padding: 10px;">
+      <h1>This is not rendered by the component. This story disappears after click (Empty your cache and/or localStorage if this page is only showing this message</h1>
+      </div>`;
+    div.appendChild(warning);
+
     return div;
   });
